# infra/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ResearchMate API
  description: Local-first ResearchMate API with swappable cloud/provider adapters.
  version: 0.1.0
paths:
...
  /api/v1/projects:
    get:
      tags:
      - projects
      summary: List Projects
      operationId: list_projects_api_v1_projects_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProjectRecord'
                type: array
                title: Response List Projects Api V1 Projects Get
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
    post:
      tags:
      - projects
      summary: Create Project
      operationId: create_project_api_v1_projects_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
  /api/v1/projects/{project_id}:
    get:
      tags:
      - projects
      summary: Get Project
      operationId: get_project_api_v1_projects__project_id__get
      security:
      - bearerAuth: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - projects
      summary: Delete Project
      operationId: delete_project_api_v1_projects__project_id__delete
      security:
      - bearerAuth: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Delete Project Api V1 Projects  Project Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/documents/upload-url:
    post:
      tags:
      - documents
      summary: Create Upload Url
      operationId: create_upload_url_api_v1_documents_upload_url_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
  /api/v1/documents:
    post:
      tags:
      - documents
      summary: Create Document
      operationId: create_document_api_v1_documents_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
  /api/v1/projects/{project_id}/documents:
    get:
      tags:
      - documents
      summary: List Project Documents
      operationId: list_project_documents_api_v1_projects__project_id__documents_get
      security:
      - bearerAuth: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentRecord'
                title: Response List Project Documents Api V1 Projects  Project Id  Documents
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/documents/{document_id}:
    get:
      tags:
      - documents
      summary: Get Document
      operationId: get_document_api_v1_documents__document_id__get
      security:
      - bearerAuth: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - documents
      summary: Delete Document
      operationId: delete_document_api_v1_documents__document_id__delete
      security:
      - bearerAuth: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Delete Document Api V1 Documents  Document Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/documents/{document_id}/complete:
    post:
      tags:
      - documents
      summary: Complete Upload
      operationId: complete_upload_api_v1_documents__document_id__complete_post
      security:
      - bearerAuth: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/UploadCompleteRequest'
              - type: 'null'
              title: Payload
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Complete Upload Api V1 Documents  Document Id  Complete
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/jobs/{job_id}:
    get:
      tags:
      - jobs
      summary: Get Job
      operationId: get_job_api_v1_jobs__job_id__get
      security:
      - bearerAuth: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/dev/traces/{trace_id}:
    get:
      tags:
      - developer-trace
      summary: Get Trace
      operationId: get_trace_api_v1_dev_traces__trace_id__get
      security:
      - bearerAuth: []
      parameters:
      - name: trace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeveloperTrace'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    DeveloperTrace:
      properties:
        trace_id:
          type: string
          format: uuid
          title: Trace Id
        user_id:
          type: string
          format: uuid
          title: User Id
        project_id:
          type: string
          format: uuid
          title: Project Id
        run_id:
          type: string
          format: uuid
          title: Run Id
        execution_plan:
          $ref: '#/components/schemas/ExecutionPlan'
        router_reason:
          type: string
          maxLength: 2000
          title: Router Reason
        retrieved_chunks:
          items:
            additionalProperties: true
            type: object
          type: array
          maxItems: 80
          title: Retrieved Chunks
        tool_calls:
          items:
            $ref: '#/components/schemas/ToolCallTrace'
          type: array
          maxItems: 40
          title: Tool Calls
        validation_result:
          additionalProperties: true
          type: object
          title: Validation Result
        latency_ms:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Latency Ms
        token_usage:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Token Usage
        errors:
          items:
            type: string
          type: array
          maxItems: 20
          title: Errors
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - trace_id
      - user_id
      - project_id
      - run_id
      - execution_plan
      - router_reason
      - validation_result
      - created_at
      title: DeveloperTrace
...
    DocumentRecord:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        project_id:
          type: string
          format: uuid
          title: Project Id
        filename:
          type: string
          title: Filename
        file_type:
          type: string
          title: File Type
        mime_type:
          type: string
          title: Mime Type
        size_bytes:
          type: integer
          title: Size Bytes
        status:
          $ref: '#/components/schemas/DocumentStatus'
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - id
      - user_id
      - project_id
      - filename
      - file_type
      - mime_type
      - size_bytes
      - status
      - created_at
      - updated_at
      title: DocumentRecord
    DocumentStatus:
      type: string
      enum:
      - uploaded
      - parsing
      - parsed
      - indexing
      - ready
      - failed
      - expired
      - deleted
      title: DocumentStatus
    ExecutionPlan:
      properties:
        source_mode:
          $ref: '#/components/schemas/SourceMode'
        task_type:
          $ref: '#/components/schemas/TaskType'
        allowed_tools:
          items:
            type: string
          type: array
          maxItems: 12
          minItems: 1
          title: Allowed Tools
        requires_local_docs:
          type: boolean
          title: Requires Local Docs
        requires_web:
          type: boolean
          title: Requires Web
        output_schema:
          type: string
          enum:
          - GroundedAnswer
          - QuizSet
          title: Output Schema
      type: object
      required:
      - source_mode
      - task_type
      - allowed_tools
      - requires_local_docs
      - requires_web
      - output_schema
      title: ExecutionPlan
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    JobRecord:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        project_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Project Id
        document_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Document Id
        type:
          type: string
          maxLength: 80
          minLength: 2
          title: Type
        status:
          $ref: '#/components/schemas/JobStatus'
        progress:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Progress
          default: 0
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - type
      - status
      - created_at
      - updated_at
      title: JobRecord
    JobStatus:
      type: string
      enum:
      - pending
      - running
      - succeeded
      - failed
      - cancelled
      title: JobStatus
    ProjectCreate:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
      type: object
      required:
      - name
      title: ProjectCreate
    ProjectRecord:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        name:
          type: string
          title: Name
        status:
          type: string
          title: Status
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - id
      - user_id
      - name
      - status
      - created_at
      - updated_at
      title: ProjectRecord
...
    SourceMode:
      type: string
      enum:
      - auto
      - local_only
      - web_only
      - hybrid
      title: SourceMode
...
    TaskType:
      type: string
      enum:
      - answer
      - quiz
      title: TaskType
    ToolCallTrace:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        tool_name:
          type: string
          maxLength: 120
          minLength: 2
          title: Tool Name
        input_summary:
          additionalProperties: true
          type: object
          title: Input Summary
        output_summary:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Output Summary
        status:
          type: string
          title: Status
        latency_ms:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Latency Ms
        error_message:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Error Message
      type: object
      required:
      - id
      - tool_name
      - input_summary
      - status
      title: ToolCallTrace
    UploadCompleteRequest:
      properties:
        checksum_sha256:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 64
          - type: 'null'
          title: Checksum Sha256
        extracted_text:
          anyOf:
          - type: string
            maxLength: 200000
          - type: 'null'
          title: Extracted Text
          description: Local development fallback. Production worker should extract
            text from R2.
      type: object
      title: UploadCompleteRequest
    UploadUrlRequest:
      properties:
        project_id:
          type: string
          format: uuid
          title: Project Id
        filename:
          type: string
          maxLength: 240
          minLength: 1
          title: Filename
        file_type:
          type: string
          enum:
          - pdf
          - docx
          - pptx
          title: File Type
        mime_type:
          type: string
          maxLength: 120
          minLength: 3
          title: Mime Type
        size_bytes:
          type: integer
          maximum: 26214400.0
          exclusiveMinimum: 0.0
          title: Size Bytes
      type: object
      required:
      - project_id
      - filename
      - file_type
      - mime_type
      - size_bytes
      title: UploadUrlRequest
    UploadUrlResponse:
      properties:
        document_id:
          type: string
          format: uuid
          title: Document Id
        upload_url:
          type: string
          maxLength: 4096
          minLength: 1
          title: Upload Url
        r2_object_key:
          type: string
          maxLength: 512
          minLength: 16
          title: R2 Object Key
        expires_in_seconds:
          type: integer
          maximum: 900.0
          minimum: 60.0
          title: Expires In Seconds
      type: object
      required:
      - document_id
      - upload_url
      - r2_object_key
      - expires_in_seconds
      title: UploadUrlResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ErrorResponse:
      $defs:
        ErrorDetail:
          properties:
            code:
              maxLength: 80
              minLength: 2
              title: Code
              type: string
            message:
              maxLength: 300
              minLength: 1
              title: Message
              type: string
            request_id:
              maxLength: 120
              minLength: 4
              title: Request Id
              type: string
          required:
          - code
          - message
          - request_id
          title: ErrorDetail
          type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      required:
      - error
      title: ErrorResponse
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or local dev token